requirements.txt
SpeechRecognition
pyttsx3
PyAudio
wikipedia
streamlit
pipwin
gTTs
google-generativeai